Foundation

Documentation

data App

The site argument for your application. This can be a good place to keep settings and values requiring initialization before your application starts running, such as database connections. Every handler will have access to the data present here.

Constructors

App

Fields

settings :: AppConfig DefaultEnv Extra

getStatic :: Static

Settings for static file serving.

connPool :: PersistConfigPool PersistConfig

Database connection pool.

httpManager :: Manager

persistConfig :: PersistConfig

data AppMessage

Constructors

MsgHello

Creates the route datatype AppRoute. Every valid URL in your

Creates the associated type:

Creates the value resourcesApp which contains information on the

type Handler = GHandler App App

type Widget = GWidget App App ()

resourcesApp :: [ResourceTree String]

type Form x = Html -> MForm App App (FormResult x, Widget)

getExtra :: Handler Extra

Get the Extra value, used to hold data from the settings.yml file.