Settings

Description

Settings are centralized, as much as possible, into this file. This includes database connection settings, static file locations, etc. In addition, you can configure a number of different aspects of Yesod by overriding methods in the Yesod typeclass. That instance is declared in the Foundation.hs file.

Documentation

type PersistConfig = SqliteConf

Which Persistent backend this site is using.

staticDir :: FilePath

The location of static files on your system. This is a file system path. The default value works properly with your scaffolded site.

staticRoot :: AppConfig DefaultEnv x -> Text

The base URL for your static files. As you can see by the default value, this can simply be static appended to your application root. A powerful optimization can be serving static files from a separate domain name. This allows you to use a web server optimized for static files, more easily set expires and cache values, and avoid possibly costly transference of cookies on static files. For more information, please see: http:code.google.comspeedpage-speeddocsrequest.html#ServeFromCookielessDomain

If you change the resource pattern for StaticR in Foundation.hs, you will have to make a corresponding change here.

To see how this value is used, see urlRenderOverride in Foundation.hs

widgetFileSettings :: WidgetFileSettings

Settings for widgetFile, such as which template languages to support and default Hamlet settings.

widgetFile :: String -> Q Exp

data Extra

Constructors

Extra

Fields

extraCopyright :: Text

extraAnalytics :: Maybe Text

Google Analytics

parseExtra :: DefaultEnv -> Object -> Parser Extra