Settings.StaticFiles

Documentation

staticSite :: IO Static

use this to create your static file serving site

img_glyphicons_halflings_png :: StaticRoute

This generates easy references to files in the static directory at compile time, giving you compile-time verification that referenced files exist. Warning: any files added to your static directory during run-time can't be accessed this way. You'll have to use their FilePath or URL to access them.

img_glyphicons_halflings_white_png :: StaticRoute

css_bootstrap_css :: StaticRoute